Update Keylogger
As the video I have attached shows, although it is a very basic tool, this keylogger could be used for a multitude of security tests. An example; a redteam could use this tool as malware, and gain access to different users account information and other sensitive user inputs. The keylogger records individual keystrokes, so while it may seem very basic and hard to comprehend, IT DOES ACKNOWLEDGE WHEN I AM HOLDING SHIFT or USING CAPS LOCK. It also picks up on     space bar     inputs and back space.